Blaze.ByteString.Builder.Char.Utf8

Description

Writes and Builders for serializing Unicode characters using the UTF-8 encoding.

Writing UTF-8 encoded characters to a buffer

writeChar :: Char -> Write

Write a UTF-8 encoded Unicode character to a buffer.

Creating Builders from UTF-8 encoded characters

fromChar :: Char -> Builder

O(1). Serialize a Unicode character using the UTF-8 encoding.

fromString :: String -> Builder

O(n). Serialize a Unicode String using the UTF-8 encoding.

fromShow :: Show a => a -> Builder

O(n). Serialize a value by Showing it and UTF-8 encoding the resulting String.

fromText :: Text -> Builder

O(n). Serialize a strict Unicode Text value using the UTF-8 encoding.

fromLazyText :: Text -> Builder

O(n). Serialize a lazy Unicode Text value using the UTF-8 encoding.