Blaze.ByteString.Builder.Char8

Description

Note: This package is intended for low-level use like implementing protocols. If you need to serialize Unicode characters use one of the UTF encodings (e.g. 'Blaze.ByteString.Builder.Char.UTF-8').

Writes and Builders for serializing the lower 8-bits of characters.

This corresponds to what the bytestring package offer in Char8.

Writing Latin-1 (ISO 8859-1) encodable characters to a buffer

writeChar :: Char -> Write

Write the lower 8-bits of a character to a buffer.

Creating Builders from Latin-1 (ISO 8859-1) encodable characters

fromChar :: Char -> Builder

O(1). Serialize the lower 8-bits of a character.

fromString :: String -> Builder

O(n). Serialize the lower 8-bits of all characters of a string

fromShow :: Show a => a -> Builder

O(n). Serialize a value by Showing it and serializing the lower 8-bits of the resulting string.

fromText :: Text -> Builder

O(n). Serialize the lower 8-bits of all characters in the strict text.

fromLazyText :: Text -> Builder

O(n). Serialize the lower 8-bits of all characters in the lazy text.