Blaze.ByteString.Builder.Html.Utf8

Description

Writes and Builders for serializing HTML escaped and UTF-8 encoded characters.

This module is used by both the 'blaze-html' and the 'hamlet' HTML templating libraries. If the Builder from 'blaze-builder' replaces the Builder implementation, this module will most likely keep its place, as it provides a set of very specialized functions.

Documentation

module Blaze.ByteString.Builder.Char.Utf8

Writing HTML escaped and UTF-8 encoded characters to a buffer

writeHtmlEscapedChar :: Char -> Write

Write a HTML escaped and UTF-8 encoded Unicode character to a bufffer.

Creating Builders from HTML escaped and UTF-8 encoded characters

fromHtmlEscapedChar :: Char -> Builder

O(1). Serialize a HTML escaped Unicode character using the UTF-8 encoding.

fromHtmlEscapedString :: String -> Builder

O(n). Serialize a HTML escaped Unicode String using the UTF-8 encoding.

fromHtmlEscapedShow :: Show a => a -> Builder

O(n). Serialize a value by Showing it and then, HTML escaping and UTF-8 encoding the resulting String.

fromHtmlEscapedText :: Text -> Builder

O(n). Serialize a HTML escaped strict Unicode Text value using the UTF-8 encoding.

fromHtmlEscapedLazyText :: Text -> Builder

O(n). Serialize a HTML escaped Unicode Text using the UTF-8 encoding.